🐙️ueli🎆️26🌙️06☀️04🚩️574
🎆️26🌙️08☀️26⏰️13⌚️46⏱️33
⚡️Mount typed-command full operations on retained worker sessions with bounded publication turns and instance operation owners across the OS plugin host
🌊️Expand flow VCS retained lifecycle with oracle fixtures and a WASM host protocol for pack-level interactive jobs
🧵️Add fixed-operation registry law fixtures and scheduler authority checks to the interactive job module
⚖️Wire verify interactivity, tool-jobs, apps, and dependencies gates into root script and launch.json with per-app dev entries
🖱️Split UI render surfaces for Metal Objective-C bindings, WebGPU browser tests, and native TUI window hosting
📺️Retire legacy plugin-runtime bridges in renderer engine elements for NodeGraph, WasmSessionLoader, and Shell host wiring
🧩Refresh puzzle and sourcing plugin editors, oracle fixtures, and mutate tests for retained interactive job contracts
🎫️Archive PHASE-1-5 de-async sweep tickets and land interactive-job migration cohort audits with demonstrator boot probes

📊️metric📃uloc💯️4.2M📈️91.7k➗️2.23➕️93.5k✏️22k➖️1.85k🟰️117k
📊️metric💾size💯️201MB📈️6.08MB➗️3.12➕️6.31MB✏️41MB➖️229KB🟰️47.5MB
📊️metric🦀️rust📃uloc💯️1.96M📈️37.3k➗️1.94➕️39.1k✏️20.1k➖️1.8k🟰️61k
📊️metric🦀️rust💾size💯️105MB📈️1.86MB➗️1.8➕️2.09MB✏️32.6MB➖️226KB🟰️35MB
📊️metric🧾️json📃uloc💯️1.53M📈️21.2k➗️1.41➕️21.2k✏️1.03k➖️8🟰️22.3k
📊️metric🧾️json💾size💯️62.8MB📈️2.45MB➗️4.05➕️2.45MB✏️4.1MB➖️1.65KB🟰️6.55MB
📊️metric🟦️typescript📃uloc💯️381k📈️3.56k➗️0.942➕️3.56k✏️670➖️5🟰️4.24k
📊️metric🟦️typescript💾size💯️20.2MB📈️223KB➗️1.12➕️224KB✏️4.05MB➖️420B🟰️4.27MB
📊️metric🔵️go📃uloc💯️106k
📊️metric🔵️go💾size💯️3.41MB
📊️metric🐍️python📃uloc💯️55.8k📈️27.1k➗️94.3➕️27.1k✏️3🟰️27.1k
📊️metric🐍️python💾size💯️2.37MB📈️1.33MB➗️128➕️1.33MB✏️19.3KB🟰️1.35MB
📊️metric🟨️javascript📃uloc💯️55.3k📈️1.2k➗️2.22➕️1.2k✏️162🟰️1.36k
📊️metric🟨️javascript💾size💯️2.17MB📈️65.4KB➗️3.11➕️65.4KB✏️60KB🟰️125KB
📊️metric🟣️csharp📃uloc💯️32.9k
📊️metric🟣️csharp💾size💯️1.44MB
📊️metric📝️markdown📃uloc💯️29.7k
📊️metric📝️markdown💾size💯️959KB
📊️metric🎨️css📃uloc💯️27.1k
📊️metric🎨️css💾size💯️685KB
📊️metric📐️tex📃uloc💯️9.45k
📊️metric📐️tex💾size💯️473KB
📊️metric⚙️toml📃uloc💯️6.4k📉️12➗️0.187➕️25✏️18➖️37🟰️80
📊️metric⚙️toml💾size💯️304KB📈️664B➗️0.219➕️1.85KB✏️96KB➖️1.18KB🟰️99.1KB
📊️metric🌐️html📃uloc💯️5.04k
📊️metric🌐️html💾size💯️308KB
📊️metric🛢️sql📃uloc💯️1.87k
📊️metric🛢️sql💾size💯️78.6KB
📊️metric📄️xml📃uloc💯️1.36k📈️1.32k➗️3060➕️1.32k🟰️1.32k
📊️metric📄️xml💾size💯️135KB📈️133KB➗️7100➕️133KB🟰️133KB
📊️metric📋️yaml📃uloc💯️1.23k
📊️metric📋️yaml💾size💯️32.7KB
📊️metric💠️powershell📃uloc💯️906
📊️metric💠️powershell💾size💯️37KB
📊️metric🐚️shell📃uloc💯️590
📊️metric🐚️shell💾size💯️21.3KB
📊️metric📑️csv📃uloc💯️174📈️51➗️41.5➕️51🟰️51
📊️metric📑️csv💾size💯️29.8KB📈️11.8KB➗️65.3➕️11.8KB🟰️11.8KB

+++ ✏️s/🔌️plugins/🗄️stdio/🗿️artifacts/🖊️dwg/🏅️standards/🔖️ac1018/🪆️subsets/✳️any/🧬️schema/🧬️mutations/📄set-snapshot/🧪️tests/bumps-the-auxiliary-save-counter/🦀️component.rs
@@ -37,7 +37,7 @@ async fn applies_to_committed_after() {
     let outcome = apply_dwg_mutation(&mut snapshot, &mutation());
     assert!(outcome.messages().is_empty(), "dwg-ac1018/set-snapshot: a genuinely changed drawing must not raise any message");
     assert_eq!(snapshot.auxiliary_header.total_saves, 8, "dwg-ac1018/set-snapshot: the save counter must advance");
-    assert_eq!(snapshot.auxiliary_header.save_generation, snapshot.auxiliary_header.terminal_save_generation, "dwg-ac1018/set-snapshot: the terminal save generation must track the current one");
+    assert_eq!(snapshot.auxiliary_header.save_generation, u32::from(snapshot.auxiliary_header.terminal_save_generation), "dwg-ac1018/set-snapshot: the terminal save generation must track the current one");
     assert_eq!(snapshot.revision_history.revisions, vec![7], "dwg-ac1018/set-snapshot: the revision history is a separate block and must not follow the save counter");
     assert_eq!(snapshot, expected_after(), "dwg-ac1018/set-snapshot: applied state differs from the committed after-snapshot");
 }
